ml/requirements.txt
Flask==3.1.1
flask_cors==6.0.1
holidays==0.76
joblib==1.4.2
pandas==2.3.1
PyJWT==2.10.1
python-dotenv==1.1.1
pytz==2025.2
Requests==2.32.4
Shapely==2.1.1
scikit-learn==1.3.2
xgboost==2.0.3
numpy==1.26.4
gunicorn==21.2.0